ssh: Ignore window adjust messages for closed channels
If a channel is closed, the ets cache has been cleared for the
ChannelId. If this is the case, cache_lookup will return undefined,
something that was unhandled by the previous code -- resulting in a
non-match error.

Both openssh and libssh2 seem to ignore the message in this situation.
(Openssh logs it, but without signalling any errors. libssh2 doesn't
even do that.)

  NOTE:

  This change might be a bandaid on some other issue in channel
  closure negotiation. Quick read of RFC 4252 is that the both send EOF
  independently of each other, and then they each send CLOSE. The
  implementation should not consider the channel closed/avaiable for
  reallocation until this happens.

  If both peers have sent CLOSE, we wouldn't expect a window adjust
  message (unless some out of order messages appear). If one or more of
  the peers haven't sent CLOSE, the channel should remain open. Do we
  close it when we send CLOSE? If so, is that a problem (assuming we
  ignore any subsequent messages)?